Output area centroids

This dataset represents geographic centroids of Output Areas in Leeds, UK.

This dataset represents geographic centroids of Output Areas in Leeds, UK.

Note

The schema data can be (re-)generated using code in the data-raw directory.

The schema data can be (re-)generated using code in the data-raw directory.

Examples

head(od_data_centroids2)

#> Simple feature collection with 6 features and 1 field
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: -1.597494 ymin: 53.74982 xmax: -1.351878 ymax: 53.90933
#> Geodetic CRS:  WGS 84
#> # A tibble: 6 x 2
#>   OA11CD                geometry
#>   <chr>              <POINT [°]>
#> 1 E00170040 (-1.532374 53.79049)
#> 2 E00058260   (-1.5588 53.84906)
#> 3 E00058975 (-1.356291 53.90933)
#> 4 E00057789  (-1.536643 53.7742)
#> 5 E00058985 (-1.351878 53.87727)
#> 6 E00058111 (-1.597494 53.74982)

head(od_data_centroids2)

#> Simple feature collection with 6 features and 1 field
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: -1.597494 ymin: 53.74982 xmax: -1.351878 ymax: 53.90933
#> Geodetic CRS:  WGS 84
#> # A tibble: 6 x 2
#>   OA11CD                geometry
#>   <chr>              <POINT [°]>
#> 1 E00170040 (-1.532374 53.79049)
#> 2 E00058260   (-1.5588 53.84906)
#> 3 E00058975 (-1.356291 53.90933)
#> 4 E00057789  (-1.536643 53.7742)
#> 5 E00058985 (-1.351878 53.87727)
#> 6 E00058111 (-1.597494 53.74982)