Origin-destination datasets

Datasets representing top commuter desire lines in Leeds based on the 2011 Census. The first two variables of the data frame are the zone code of origin and destination, respectively. The other columns record the number of people who travel by different modes, including all, train, bus, bicycle and by foot.

Details

od_data_df_medium is a larger dataset with the same variables, with around 10k rows.

Note

The schema data can be (re-)generated using code in the data-raw directory.

Examples

od_data_df
#>   geo_code1 geo_code2  all train bus taxi car_driver car_passenger bicycle foot
#> 1 E02002384 E02006875  966    14 153   14         69            18      13  679
#> 2 E02002404 E02006875 1145     6 174   17         96            38      10  798
#> 3 E02006875 E02006875 1791    21  38    5         69             7       8 1637
#> 4 E02006876 E02006875 1035    11 132    6         97            24      10  749
#> 5 E02006861 E02002392  453     1  51    0         51             6      26  317
#> 6 E02006875 E02002392  286     2  15    5         16             2      10  235
#> 7 E02002392 E02006875  753    10  91   21         33             7      19  571