Origin-destination data with destinations in a different layer than origins

This dataset represents commuter flows between Output Areas and Workplace Zones, the most detailed open OD data in the UK. See https://wicid.ukdataservice.ac.uk/ and the script data-raw/od_wpz.R in the od package's GitHub repo.

Details

The dataset reports (in the 3rd column) the number of people travelling between origins and destinations.

Note

The schema data can be (re-)generated using code in the data-raw directory.

Examples

head(od_data_df2)

#> # A tibble: 6 x 3
#>   o         d             n
#>   <chr>     <chr>     <dbl>
#> 1 E00056754 E33009626     2
#> 2 E00056914 E33012372     1
#> 3 E00056966 E33012253     1
#> 4 E00057598 E33012219     1
#> 5 E00057598 E33012368     2
#> 6 E00057598 E33012474     1