Convert a series of points into a dataframe of origins and destinations

Takes a series of geographical points and converts them into a data.frame representing the potential flows, or 'spatial interaction', between every combination of points.

points_to_od(p, interzone_only = FALSE)

Arguments

p
A spatial points object

interzone_only
Should the result only include interzonal OD pairs, in which the ID of the origin is different from the ID of the destination zone? `FALSE` by default

Examples

p = od_data_centroids[1:3, ]
(od1 = points_to_od(p))

#>           O         D
#> 1 E02002407 E02002407
#> 2 E02002407 E02002336
#> 3 E02002407 E02002353
#> 4 E02002336 E02002407
#> 5 E02002336 E02002336
#> 6 E02002336 E02002353
#> 7 E02002353 E02002407
#> 8 E02002353 E02002336
#> 9 E02002353 E02002353

points_to_od(p, interzone_only = TRUE)

#>           O         D
#> 2 E02002407 E02002336
#> 3 E02002407 E02002353
#> 4 E02002336 E02002407
#> 6 E02002336 E02002353
#> 7 E02002353 E02002407
#> 8 E02002353 E02002336

od1

#>           O         D
#> 1 E02002407 E02002407
#> 2 E02002407 E02002336
#> 3 E02002407 E02002353
#> 4 E02002336 E02002407
#> 5 E02002336 E02002336
#> 6 E02002336 E02002353
#> 7 E02002353 E02002407
#> 8 E02002353 E02002336
#> 9 E02002353 E02002353

od1$v = 1
unique(od1)

#>           O         D v
#> 1 E02002407 E02002407 1
#> 2 E02002407 E02002336 1
#> 3 E02002407 E02002353 1
#> 4 E02002336 E02002407 1
#> 5 E02002336 E02002336 1
#> 6 E02002336 E02002353 1
#> 7 E02002353 E02002407 1
#> 8 E02002353 E02002336 1
#> 9 E02002353 E02002353 1

od_oneway(od1)

#>           o         d v
#> 1 E02002336 E02002336 1
#> 2 E02002336 E02002353 2
#> 3 E02002353 E02002353 1
#> 4 E02002336 E02002407 2
#> 5 E02002353 E02002407 2
#> 6 E02002407 E02002407 1