Create matrices representing origin-destination coordinates

This function takes a wide range of input data types (spatial lines, points or text strings) and returns a matrix of coordinates representing origin (ox, oy) and destination (dx, dy) points.

od_coordinates(x, p = NULL, verbose = FALSE, sfnames = FALSE)

Arguments

x
A data frame in which the first two columns are codes representing points/zones of origin and destination

p
Points representing origins and destinations

verbose
Print messages providing progress updates? FALSE by default

sfnames
Should output column names be compatible with the sf package?

Examples

x = od_data_df
p = od_data_centroids
od_coordinates(x, p)[1:2, ]

#>          ox       oy        dx       dy
#> 6 -1.534957 53.82859 -1.545708 53.79593
#> 7 -1.581773 53.81860 -1.545708 53.79593

od_coordinates(x, p, sfnames = TRUE)[1:2, ]

#>           x        y         x        y
#> 6 -1.534957 53.82859 -1.545708 53.79593
#> 7 -1.581773 53.81860 -1.545708 53.79593

od_coordinates(x, p, verbose = TRUE)[1:2, ]

#>  points not in od data removed.

#>          ox       oy        dx       dy
#> 6 -1.534957 53.82859 -1.545708 53.79593
#> 7 -1.581773 53.81860 -1.545708 53.79593

# x[[1]][1] =  "404"
# Next line will error:
# od_coordinates(x, p, verbose = TRUE)[1:2, ]
# From original stplanr function:
# od_coords(from = c(0, 52), to = c(1, 53)) # lon/lat coordinates
# od_coords(from = cents[1, ], to = cents[2, ]) # Spatial points
# od_coords(cents_sf[1:3, ], cents_sf[2:4, ]) # sf points
# # od_coords("Hereford", "Leeds") # geocode locations
# od_coords(flowlines[1:3, ])
# od_coords(flowlines_sf[1:3, ])