Filter OD datasets

This function takes and OD dataset and a character vector of codes and returns an OD dataset with rows matching origin and destinations zones present in the codes.

od_filter(x, codes, silent = FALSE)

Arguments

x
A data frame in which the first two columns are codes representing points/zones of origin and destination

codes
The zone codes that must be in origins and destination

silent
Hide messages? FALSE by default.

Value

A data frame

Examples

x = od_data_df
z = od_data_zones
codes = z[[1]]
z_in_x_o = codes %in% x[[1]]
z_in_x_d = codes %in% x[[2]]
sum(z_in_x_d)

#> [1] 2

sum(z_in_x_o)

#> [1] 5

z = z[which(z_in_x_o | z_in_x_d)[-1], ]
z[[1]]

#> [1] "E02002373" "E02002385" "E02002392" "E02006852" "E02006861" "E02006875"

unique(c(x[[1]], x[[2]]))

#> [1] "E02002363" "E02002373" "E02002385" "E02006852" "E02006861" "E02006875"
#> [7] "E02002392"

try(od_to_sf(x, z)) # fails

#> 1 origins with no match in zone ids

#> 0 destinations with no match in zone ids

#>  points not in od data removed.

#> Simple feature collection with 5 features and 10 fields
#> geometry type:  LINESTRING
#> dimension:      XY
#> bbox:           xmin: -1.586482 ymin: 53.79587 xmax: -1.546024 ymax: 53.83316
#> z_range:        zmin: NA zmax: NA
#> m_range:        mmin: NA mmax: NA
#> geographic CRS: WGS 84
#>   geo_code1 geo_code2  all train bus taxi car_driver car_passenger bicycle foot
#> 1 E02002373 E02006875 1037   111 424   20        155            30      73  214
#> 2 E02002385 E02006875  958   121 334   19        118            25      52  283
#> 3 E02006852 E02002392  437     1  96    3        142            26      61  108
#> 4 E02006852 E02006875 1221    14 509   13        401            50      99  118
#> 5 E02006861 E02006875 1177    43 400   30        123            28      56  492
#>                         geometry
#> 1 LINESTRING (-1.581617 53.81...
#> 2 LINESTRING (-1.582003 53.80...
#> 3 LINESTRING (-1.586482 53.83...
#> 4 LINESTRING (-1.586482 53.83...
#> 5 LINESTRING (-1.564756 53.81...

nrow(x)

#> [1] 6

x = od_filter(x, z[[1]])

#> 1 origins with no match in zone ids

#> 0 destinations with no match in zone ids

nrow(x)

#> [1] 5

od_to_sf(x, z)

#> 0 origins with no match in zone ids

#> 0 destinations with no match in zone ids

#>  points not in od data removed.

#> Simple feature collection with 5 features and 10 fields
#> geometry type:  LINESTRING
#> dimension:      XY
#> bbox:           xmin: -1.586482 ymin: 53.79587 xmax: -1.546024 ymax: 53.83316
#> z_range:        zmin: NA zmax: NA
#> m_range:        mmin: NA mmax: NA
#> geographic CRS: WGS 84
#>   geo_code1 geo_code2  all train bus taxi car_driver car_passenger bicycle foot
#> 1 E02002373 E02006875 1037   111 424   20        155            30      73  214
#> 2 E02002385 E02006875  958   121 334   19        118            25      52  283
#> 3 E02006852 E02002392  437     1  96    3        142            26      61  108
#> 4 E02006852 E02006875 1221    14 509   13        401            50      99  118
#> 5 E02006861 E02006875 1177    43 400   30        123            28      56  492
#>                         geometry
#> 1 LINESTRING (-1.581617 53.81...
#> 2 LINESTRING (-1.582003 53.80...
#> 3 LINESTRING (-1.586482 53.83...
#> 4 LINESTRING (-1.586482 53.83...
#> 5 LINESTRING (-1.564756 53.81...