Combine two ID values to create a single ID number

Combine two ID values to create a single ID number

od_id_szudzik(x, y, ordermatters = FALSE)

od_id_max_min(x, y)

od_id_character(x, y)

Arguments

x
a vector of numeric, character, or factor values

y
a vector of numeric, character, or factor values

ordermatters
logical, does the order of values matter to pairing, default = FALSE

Details

In OD data it is common to have many 'oneway' flows from "A to B" and "B to A". It can be useful to group these an have a single ID that represents pairs of IDs with or without directionality, so they contain 'twoway' or bi-directional values.

od_id* functions take two vectors of equal length and return a vector of IDs, which are unique for each combination but the same for twoway flows.

the Szudzik pairing function, on two vectors of equal length. It returns a vector of ID numbers.

This function superseeds od_id_order as it is faster on large datasets

See also

Examples

(d = od_data_df[2:9, 1:2])

#>      geo_code1 geo_code2
#> 2    E02002373 E02006875
#> 3    E02002385 E02006875
#> 4    E02006852 E02002392
#> 5    E02006852 E02006875
#> 6    E02006861 E02006875
#> NA        <NA>      <NA>
#> NA.1      <NA>      <NA>
#> NA.2      <NA>      <NA>

(id = od_id_character(d[[1]], d[[2]]))

#> [1] "E02002373 E02006875" "E02002385 E02006875" "E02002392 E02006852"
#> [4] "E02006852 E02006875" "E02006861 E02006875" "NA NA"              
#> [7] "NA NA"               "NA NA"              

duplicated(id)

#> [1] FALSE FALSE FALSE FALSE FALSE FALSE  TRUE  TRUE

od_id_szudzik(d[[1]], d[[2]])

#> [1] 26 27 39 28 29 NA NA NA

od_id_max_min(d[[1]], d[[2]])

#> [1] 16 17 24 18 19 NA NA NA

n = 100
ids = as.character(runif(n, 1e4, 1e7 - 1))
# benchmark of methods:
x = data.frame(
  id1 = rep(ids, times = n),
  id2 = rep(ids, each = n),
  val = 1,
  stringsAsFactors = FALSE
)
bench::mark(
  check = FALSE, iterations = 10,
  od_id_order(x),
  od_id_character(x$id1, x$id2),
  od_id_szudzik(x$id1, x$id2),
  od_id_max_min(x$id1, x$id2)
)

#> # A tibble: 4 x 13
#>   expression                       min median `itr/sec` mem_alloc `gc/sec` n_itr
#>   <bch:expr>                    <bch:> <bch:>     <dbl> <bch:byt>    <dbl> <int>
#> 1 od_id_order(x)                4.99ms 5.73ms      177.  240.31KB     19.7     9
#> 2 od_id_character(x$id1, x$id2) 5.24ms 5.83ms      140.  234.52KB      0      10
#> 3 od_id_szudzik(x$id1, x$id2)   1.32ms 1.73ms      586.    1.64MB      0      10
#> 4 od_id_max_min(x$id1, x$id2)   1.14ms 1.29ms      775.  974.84KB      0      10
#> # … with 6 more variables: n_gc <dbl>, total_time <bch:tm>, result <list>,
#> #   memory <list>, time <list>, gc <list>